factor out fit result into own module
